+++ back-end/database/repository/book_repository.js
@@ -24,13 +24,40 @@ class BookRepository {
     }
   }
 
+  async GetAllBooks() {
+    try {
+      const books = await BookModel.find().populate("author").exec(); //return author document with collection
+      //check avilability of book and error handling
+      if (books.length === 0) {
+        return {
+          error: true,
+          result: "Not found any books!",
+        };
+      }
+
+      return { error: false, result: books };
+    } catch (err) {
+      throw APIError(
+        "API Error",
+        STATUS_CODES.INTERNAL_ERROR,
+        "Invalid operations!"
+      );
+    }
+  }
+
   async IsBooksAlreadyRegistered({ name, isbn }) {
     try {
-      const existingBook = await BookModel.findOne({
+      const existingBookWithName = await BookModel.find({
         name,
+      });
+      const existingBookWithIsbn = await BookModel.find({
         isbn,
       });
-      return existingBook;
+      //Validating book name and ISBN
+      return existingBookWithName.length === 0 &&
+        existingBookWithIsbn.length === 0
+        ? false
+        : true;
     } catch (err) {
       throw APIError("API Error", STATUS_CODES.INTERNAL_ERROR, "Failed!");
     }