+++ back-end/utils/utils.js
@@ -1,7 +1,14 @@
 export const FormateData = (data) => {
-    if (data) {
-      return { data };
-    } else {
-      throw new Error("Data Not found!");
-    }
-  };
+  if (data) {
+    return { data };
+  } else {
+    throw new Error("Data Not found!");
+  }
+};
+
+export const ValidateIsbn = (data) => {
+  const regex =
+    "^(?:ISBN(?:-10)?:? )?(?=[0-9X]{10}$|(?=(?:[0-9]+[- ]){3})[- 0-9X]{13}$)[0-9]{1,5}[- ]?[0-9]+[- ]?[0-9]+[- ]?[0-9X]$";
+  const isbnReg = new RegExp(regex);
+  return isbnReg.test(data);
+};